Use BGR for color images
See #59. Caffe uses OpenCV to read encoded images from LMDB, which
results in a caffemodel that expects images in BGR, not RGB.
So, I'm standardizing on images within DIGITS to match.

This is not the preferred solution - this requires all other frameworks
that consume DIGITS datasets to convert from BGR to RGB. But it is the
simplest and fastest solution.

* Implemented upgrade paths for previously created datasets and models
* Still need to add tests for this